TypeScript and project changes
