+++ .github/workflows/main.yml
@@ -2,19 +2,18 @@ name: Node.js CI
 
 on:
   push:
-    branches: [ master ]
+    branches: [master]
   pull_request:
-    branches: [ master ]
+    branches: [master]
 
 jobs:
   build:
-
     # runs-on: ubuntu-latest # node 11 fails with Python 2 exception for print
-    runs-on: ubuntu-20.04 
+    runs-on: ubuntu-20.04
 
     strategy:
       matrix:
-        node-version: [11.x, 12.x, 14.x, 16.x]
+        node-version: [11.x, 12.x, 14.x, 16.x, 22.x]
 
     steps:
       - uses: actions/checkout@v3
@@ -30,7 +29,6 @@ jobs:
       # - run: npm test
 
   verify-json-validation:
-    
     runs-on: ubuntu-latest
 
     strategy:
@@ -42,5 +40,5 @@ jobs:
       - name: Validate JSON ${{ matrix.file-name }}.json
         uses: docker://orrosenblatt/validate-json-action:latest
         env:
-          INPUT_SCHEMA: /test/${{ matrix.file-name }}_schema.json
+          INPUT_SCHEMA: /tests/schemas/${{ matrix.file-name }}_schema.json
           INPUT_JSONS: /server/json/${{ matrix.file-name }}.json
