--- devs.txt
@@ -1,1 +1,0 @@
-We really need some types for the node-ld library.
