+++ src/enums/Emits.ts
@@ -5,4 +5,5 @@ enum Emits {
   FadeOne = "Fade One",
   FadeAll = "Fade All",
   ConnectionAffirmation = "Connection True",
+  ConnectionDenial = "Connection False",
 }
