+++ src/utils/toytags.ts
@@ -3,44 +3,6 @@ const path = require("path");
 
 const toytagsPath = path.join(__dirname, "server/json/toytags.json");
 
-//This searches toytags.json and returns to UID of the entry with the matching index.
-export function getUIDFromIndex(index: number) {
-  const data = internal_get();
-  for (let i = 0; i < data.length; i++) {
-    const entry = data[i];
-
-    if (entry.index == index) {
-      return entry.uid;
-    }
-  }
-}
-
-export function getEntryFromUID(uid: string) {
-  const data = internal_get();
-
-  for (let i = 0; i < data.length; i++) {
-    const entry = data[i];
-
-    if (entry.uid == uid) {
-      return entry;
-    }
-  }
-}
-//This updates the pad index of a tag in toytags.json, so that info can be accessed locally.
-export function updatePadIndex(uid: string, index: number) {
-  const data = internal_get();
-
-  for (let i = 0; i < data.length; i++) {
-    const entry = data[i];
-
-    if (entry.uid == uid) {
-      entry.index = index;
-      break;
-    }
-  }
-
-  internal_write(data);
-}
 //This updates the provided datatype, of the entry with the matching uid, with the provided data.
 export function updateKey(
   uid: string,
@@ -63,7 +25,7 @@ export function updateKey(
 
 export function updateKeys(
   uid: string,
-  bundle: [{ key: keyof Toytag; value: Toytag[keyof Toytag] }]
+  bundle: { key: string; value: string }[]
 ) {
   const data = internal_get();
 
@@ -71,11 +33,9 @@ export function updateKeys(
     const entry = data[i];
 
     if (entry.uid == uid) {
-      bundle.forEach(
-        (data: { key: keyof Toytag; value: Toytag[keyof Toytag] }) => {
-          entry[data.key] = data.value;
-        }
-      );
+      bundle.forEach((data) => {
+        entry[data.key] = data.value;
+      });
       break;
     }
   }
@@ -99,16 +59,35 @@ export function unplaceAll() {
 
   internal_write(data);
 }
-export function deleteByUID(uid: string) {
+export function deleteEntry(key: keyof Toytag, value: Toytag[keyof Toytag]) {
   const data = internal_get();
 
+  let wasDeleted = false;
   for (let i = 0; i < data.length; i++) {
-    if (data[i].uid == uid) {
+    if (data[i][key] === value) {
       data.splice(i, 1);
+      wasDeleted = true;
       break;
     }
   }
-  internal_write(data);
+  if (wasDeleted) {
+    internal_write(data);
+    return true;
+  }
+  return false;
+}
+export function select(
+  key: keyof Toytag,
+  value: Toytag[keyof Toytag]
+): Toytag | null {
+  const data = internal_get();
+
+  for (let i = 0; i < data.length; i++) {
+    if (data[i][key] === value) {
+      return data[i];
+    }
+  }
+  return null;
 }
 //** Internal private calls */
 function internal_get() {
