+++ tests/schemas/charactermap_schema.json
@@ -1,0 +1,30 @@
+{
+    "type": "array",
+    "title": "Root Schema",
+    "minItems": 1,
+    "uniqueItems": true,
+    "items": {
+        "type": "object",
+        "title": "A Schema",
+        "minProperties": 4,        
+        "additionalProperties": false,
+        "properties": {
+            "id": {
+                "type": "integer",
+                "title": "The id Schema"
+            },
+            "name": {
+                "type": "string",
+                "title": "The name Schema"
+            },
+            "world": {
+                "type": "string",
+                "title": "The world Schema"
+            },
+            "abilities": {
+                "type": "string",
+                "title": "The abilities Schema"
+            }
+        }
+    }
+}
