+++ tsconfig.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
     "target": "ES2024",
-    "module": "ESNext",
+    "module": "CommonJS",
     "lib": ["DOM", "ES2024"],
     "strict": true,
     "esModuleInterop": true,
